Corrige l’échec du CI lié à libpqxx
Le CI échouait à cause d’une incompatibilité entre la version
de libpqxx installée et le compilateur utilisé.

Solution :
- build libpqxx depuis les sources avec g++-14 pour assurer la compatibilité
- utilise aussi g++-14 pour compiler googletest dans le CI
- corrige un warning de conversion dans `test_database.cpp`

+++ backend/tests/test_database.cpp
@@ -98,7 +98,7 @@ TEST_F(DatabaseTest, GetPricesNoDatesReturnsAll) {
 // sans passer par parse_date().
 TEST(DateIntegerConversionTest, RoundTripIsExact) {
     Date expected = *parse_date("2024-01-15");
-    int  days_since_epoch = expected.time_since_epoch().count();
+    int  days_since_epoch = static_cast<int>(expected.time_since_epoch().count());
     Date reconstructed{std::chrono::days{days_since_epoch}};
     EXPECT_EQ(expected, reconstructed);
 }
